requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
et-xmlfile==1.1.0
Flask==2.1.2
fpdf==1.7.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.4
openpyxl==3.2.0b1
pandas==1.4.2
pycparser==2.21
pypng==0.0.21
PyQRCode==1.2.1
python-dateutil==2.8.2
python-firebase @ git+https://github.com/ozgur/python-firebase@0d79d7609844569ea1cec4ac71cb9038e834c355
pytz==2022.1
requests==2.27.1
six==1.16.0
urllib3==1.26.9
Werkzeug==2.1.2
xlrd==2.0.1
zipp==3.8.0
